==> notebooks/1_electrical_electronic_equipment.ipynb <==
# 1. Electrical & Electronic Equipment (EEE)

This dataset includes the description provided by the United Nations University (UNU) about the type of electronic/electrical equipment (EEE), the year of introduction of EEE into the market and the average weight of EEE over the years. (Source: https://github.com/Statistics-Netherlands/ewaste)

Relevant variables include:

1. **UNU_Key**: keys referenced to EEE items description as defined by the United Nations University (UNU)
2. **Description**: description associated to the UNU_Key for EEE.  
3. **IntroductionYear**: the year of introduction of EEE into the market.
4. **AverageWeight**: average weight of a given EEE item in a given year.
5. **Year**: year associated to the average weight of a given EEE item.

A new data frame is created from the raw data. The relevant variables listed above are analyzed and cleaned when necessary. The columns are renamed to avoid conlficts with other data frames. The final data frame is finally saved into a csv file.

### Data loading and cleaning

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

#### 1. Description of EEE keys 

# Import raw data

description = pd.read_csv("../data/raw/1_htbl_Key_Description.csv")

# Rename columns for easier handling

description = description.rename(columns={"UNU_Key": "eee_key", "Description": "eee_description"})

description.dtypes

# Check for missing values

description.isnull().sum()

# Delete eee_key entries that have alpha characters

description = description[~description["eee_key"].str.contains("[A-Za-z]")]

description

# Locate entries with "phone" in eee_description 

description.loc[description["eee_description"].str.contains("phone")]

# We can see that "mobile phones" are associated to eee_key = 306

# Save to csv file

description.to_csv("../Data/clean_data/1_eee_description.csv", index=False)
