==> notebooks/2_e_waste.ipynb <==
# 2. e-waste data

This dataset includes the amount of electronic/electrical equipment waste. (Source: https://github.com/Statistics-Netherlands/ewaste) Waste Electrical & Electronic Equipment (WEEE) will be referred as ***e-waste*** for simplicity. 

Relevant variables include:

1. **Country**: three-letter code of each country
2. **UNU_Key**: keys referenced to electronic items description as defined by the United Nations University (UNU)
3. **Year**: year for the e-waste entry
4. **WEEE_t**: amount of e-waste generated in the country, in tonnes 
5. **WEEE_pieces**: amount of e-waste generated in the country, in pieces count
6. **Inhabitants**: total population of a given country in a given year
7. **kpi**: amount of e-waste per inhabitant, in kg per inhabitant
8. **ppi**: amount of e-waste per inhabitant, in pieces per inhabitant


A new data frame is created from the raw data. The relevant variables listed above are analyzed and cleaned when necessary. The columns are renamed to avoid conlficts with other data frames. The final data frame is finally saved into a csv file.

### Data loading and cleaning

import pandas as pd
import numpy as np

#### 1. e-waste in Europe 
In detail, from 1980 to 2021.

# Import raw data

ewaste = pd.read_csv("../data/raw/2_tbl_WEEE.csv")

# Transform column names into lower case for easier handling

ewaste.columns = map(str.lower, ewaste.columns)

# Rename kpi and ppi columns to avoid conflict with data for POM

ewaste = ewaste.rename(columns={"unu_key": "eee_key",
                                "weee_t": "ewaste_t",
                                "weee_pieces": "ewaste_pieces",
                                "kpi": "ewaste_kpi", 
                                "ppi": "ewaste_ppi"})

# Drop useless columns

ewaste = ewaste.drop(["stratum"], axis=1)

ewaste.dtypes

# Transform unu_key and year to string object

ewaste["eee_key"] = ewaste["eee_key"].astype(str)
ewaste["year"] = ewaste["year"].astype(str)

ewaste.dtypes

# Check for missing values

ewaste.isnull().sum()

# there are 10584 missing values in inhabitants, kpi and ppi

ewaste[ewaste["inhabitants"].isnull()]["year"].value_counts()

# checking the data, we can see that population estimates from 2022 on 
# are not available

ewaste["country"].nunique()

# There are 28 countries, which are the countries of the EU

# Verify the data for each country

pd.Series(ewaste["country"]).value_counts()

ewaste.head()

# Save the data frame to a csv file

ewaste.to_csv("../Data/clean_data/2_ewaste_generation.csv", index=False)

#### 2. Extra data: ewaste worldwide from 2000 to 2021 

Total values only. btained from Kees Balde @ United Nations University.

# Import raw data

ewaste_world = pd.read_csv("../data/raw/3_extra_ewaste_world.csv")

ewaste_world.head()

ewaste_world.isnull().sum()

ewaste_world.dtypes

ewaste_world["year"] = ewaste_world["year"].astype(str)

ewaste_world.nunique()

# Save the data frame to a csv file

ewaste_world.to_csv("../Data/clean_data/3_ewaste_world_generation.csv", index=False)
