==> notebooks/3_gdp_countries.ipynb <==
# 3. GDP and countries

This dataset includes the GDP per country. (Source: https://data.worldbank.org/indicator/NY.GDP.MKTP.CD)

Relevant variables include:

1. **Country Name**: country name
2. **Country Code**: three-letter code of each country
3. **years**: year for the GDP entry

A new data frame is created from the raw data. The relevant variables listed above are analyzed and cleaned when necessary. The columns are renamed to avoid conlficts with other data frames. The final data frame is finally saved into a csv file.

### Data loading and cleaning

import pandas as pd
import numpy as np

#### 1. GDP (total per country)

# Import raw data

gdp = pd.read_csv("../data/raw/4_gdp_countries.csv")

gdp.head()

# Rename Country Name and Country Code columns to match with data for EEE and POM
# Drop useless columns

gdp = gdp.rename(columns={"Country Name": "country_name",
                          "Country Code": "country"}).drop(["Indicator Name", "Indicator Code"], axis=1)

# Create a single variable "year" and assign the value of "gdp" to it
# for each country

gdp_transform = gdp.melt(["country_name", "country"])

# Rename columns from the melt

gdp_transform = gdp_transform.rename(columns={"variable": "year", "value": "gdp"})

gdp_transform.dtypes

gdp_transform.head()

# check for missing values

gdp_transform.isna().sum()

gdp_transform.to_csv("../Data/clean_data/4_gdp_total_countries.csv", index=False)

#### 2. GDP (per capita)

# Import raw data

gdp_pc = pd.read_csv("../data/raw/5_gdp_percapita.csv")

gdp_pc.head()

# Rename Country Name and Country Code columns to match with data for EEE and POM
# Drop useless columns

gdp_pc = gdp_pc.rename(columns={"Country Name": "country_name",
                                "Country Code": "country"}).drop(["Indicator Name", "Indicator Code", ], axis=1)

# Create a single variable "year" and assign the value of "gdp" to it
# for each country

gdp_pc_transform = gdp_pc.melt(["country_name", "country"])

# Rename columns from the melt

gdp_pc_transform = gdp_pc_transform.rename(columns={"variable": "year", "value": "gdp_pc"})

gdp_pc_transform.head()

gdp_pc_transform.to_csv("../Data/clean_data/5_gdp_per_capita.csv", index=False)

#### 3. Countries aggregates

countries = pd.read_csv("../data/raw/6_countries_income_groups.csv")

countries

countries = countries.rename(columns={"Country Code": "country", 
                                      "IncomeGroup": "income_group", 
                                      "TableName": "country_name"}).drop(["Region"], axis=1)

#### Create column "region" with "europe" for the 28 EU countries, and "other" for the rest of the countries

# Create a classification for the 28 EU countries

eu_countries = ['AUT', 'BEL', 'BGR', 'CYP', 'CZE', 'DEU', 'DNK', 'ESP', 'EST',
                'FIN', 'FRA', 'GBR', 'GRC', 'HRV', 'HUN', 'IRL', 'ITA', 'LTU',
                'LUX', 'LVA', 'MLT', 'NLD', 'POL', 'PRT', 'ROU', 'SVK', 'SVN',
                'SWE']

# Create a function with loops though the list of EU countries

def match_country(country):
    k = (["Europe" for pays in eu_countries if pays in country])
    return k[0] if k else "Other"

# Pass a lambda function to create the "region" column

countries['region'] = countries.apply(lambda x: match_country(x['country']), axis=1)

countries.dtypes

countries.head()

# Locate entries with continent = europe

countries.loc[countries["region"].str.contains("Europe")].nunique()

countries.to_csv("../Data/clean_data/6_countries.csv", index=False)
