==> notebooks/5_e-waste_analysis_eu.ipynb <==
# 5. Analysis of e-waste generation in the EU

This notebook contains the analysis of the data regarding the e-weaste generated in the European Union from 1980 to 2021. A new data frame is created from the clean data. 

Relevant variables include:

1. E-waste description
2. E-waste generation in the EU
3. GDP per country
4. GDP per country (per capita)
5. E-waste categories


import pandas as pd
import numpy as np
import pymysql
import getpass
import matplotlib
import matplotlib.pyplot as plt
import mysql_connection as mc
%matplotlib inline

description = pd.read_csv("../data/clean_data/1_eee_description.csv")

ewaste = pd.read_csv("../data/clean_data/2_ewaste_generation.csv")

gdp_t = pd.read_csv("../data/clean_data/4_gdp_total_countries.csv")

gdp_pc = pd.read_csv("../data/clean_data/5_gdp_per_capita.csv")

countries = pd.read_csv("../data/clean_data/6_countries.csv")

cat = pd.read_csv("../data/clean_data/7_ewaste_category.csv")

df = description.merge(right=cat, 
                       how="left", 
                       on="category")[["eee_key", 
                                       "eee_description", 
                                       "type"]]

df2 = ewaste.merge(right=df, 
                   how="left", 
                   on="eee_key")[["year",
                                  "country",
                                  "inhabitants",
                                  "ewaste_t",
                                  "ewaste_pieces",
                                  "ewaste_kpi",
                                  "ewaste_ppi",
                                  "eee_key",
                                  "eee_description",
                                  "type"]]

df3 = df2.merge(right=countries, 
                  how="left", 
                  on="country")[["country_name",
                                 "year",
                                 "income_group", 
                                 "inhabitants",
                                 "ewaste_t", 
                                 "ewaste_pieces", 
                                 "ewaste_kpi", 
                                 "ewaste_ppi", 
                                 "eee_key", 
                                 "eee_description",
                                 "type"]]

df4 = df3.merge(right=gdp_t, 
                  how="left", 
                  on=["country_name", "year"])[["country_name",
                                                "year",
                                                "income_group",
                                                "inhabitants",
                                                "ewaste_t",
                                                "ewaste_pieces",
                                                "ewaste_kpi",
                                                "ewaste_ppi",
                                                "gdp",
                                                "eee_key",
                                                "eee_description",
                                                "type"]]

df5 = df4.merge(right=gdp_pc, 
                  how="left", 
                  on=["country_name", "year"])[["country_name",
                                                "year",
                                                "income_group",
                                                "inhabitants",
                                                "ewaste_t",
                                                "ewaste_pieces",
                                                "ewaste_kpi",
                                                "ewaste_ppi",
                                                "gdp", 
                                                "gdp_pc",
                                                "eee_key",
                                                "eee_description",
                                                "type"]]

df5.head()

# Create a column of e-waste generated in million tonnes 

df5["ewaste_mt"] = df5["ewaste_t"] / 1000000

# Create a column with e-waste equivalent to Eiffel Towers

eiffel_tower = 10100
df5["ewaste_eff"] = df5["ewaste_mt"] * 1000000 / eiffel_tower

# Let's look at the waste generated per year in the whole region

year_mt = (df5
           .groupby("year")
           .agg({"ewaste_mt": "sum", "ewaste_eff": "sum"})
           .reset_index())

year_mt

# Let's look at the e-waste equivalent to Eiffel Towers generated in a single year

def e_waste_equiv_et(year):
    print("In " + str(year) + ", " + str(year_mt.loc[year_mt["year"] == year]["ewaste_mt"].iloc[0]) 
          + " million metric tonnes of e-waste was generated. This is equivalent to " 
          + str(year_mt.loc[year_mt["year"] == year]["ewaste_eff"].iloc[0]) + " Eiffel Towers!")

e_waste_equiv_et(2000)
e_waste_equiv_et(2020)

# Let's look at the numbers per category in a given year

ewaste_cat = (df5.loc[df5["year"] == 2019]
              .groupby("eee_key")
              .agg({"ewaste_mt": "sum", "ewaste_eff": "sum", "ewaste_pieces": "sum"})
              .reset_index()).merge(right=df, 
                                    how="left", 
                                    on="eee_key")[["eee_key", 
                                                   "eee_description", 
                                                   "ewaste_mt", 
                                                   "ewaste_eff", 
                                                   "ewaste_pieces"]]

ewaste_cat.sort_values(by=["ewaste_mt"])

# Let's look at the numbers in 2019

ewaste_cat_2019 = (df5.loc[df5["year"] == 2019]
              .groupby("eee_key")
              .agg({"ewaste_mt": "sum", "ewaste_eff": "sum", "ewaste_pieces": "sum"})
              .reset_index()).merge(right=df, 
                                    how="left", 
                                    on="eee_key")[["eee_key", 
                                                   "eee_description", 
                                                   "ewaste_mt", 
                                                   "ewaste_eff", 
                                                   "ewaste_pieces"]]

ewaste_cat_2019.sort_values(by=["ewaste_mt"])

# Let's look at the numbers in Germany in 2019

(df5.loc[df5["country_name"] == "Germany"]).loc[df5["year"] == 2019][["eee_key", 
                           "eee_description", 
                           "ewaste_mt", 
                           "ewaste_eff", 
                           "ewaste_pieces"]].sort_values(by=["ewaste_mt"], ascending=False)

ewaste_cat_1

# Let's look at the numbers per type (combined categories) in a given year

ewaste_type = (df5
               .loc[df5["year"] == 2019]
               .groupby("type")
               .agg({"ewaste_mt": "sum", "ewaste_eff": "sum"})
               .reset_index())

ewaste_type.sort_values(by=["ewaste_mt"])

# Let's look at the numbers per country in a given year

ewaste_countries = (df5
                    .groupby(["country_name", "year"])
                    .agg({"ewaste_mt": "sum", "ewaste_eff": "sum"})
                    .reset_index())

ewaste_countries

# Let's look at the numbers per country in a given year

(ewaste_countries[ewaste_countries["country_name"] == "Germany"])

# df5.to_csv("../Data/clean_data/10_ewaste_eu.csv", index=False)
